pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Nicolas Bigot"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pydantic = "^2.8.2"
pyjwt = "^2.8.0"
requests = "^2.32.3"
jinja2 = "^3.1.4"
python-dateutil = "^2.9.0.post0"
pyyaml = "^6.0.1"
jsonschema = "^4.23.0"
python-json-logger = "^2.0.7"
boto3 = "^1.34.151"
pandas = "^2.2.2"
psycopg2 = "^2.9.9"
mysql = "^0.0.3"
redis = "^5.0.8"
pika = "^1.3.2"
pytest = "^8.3.2"
mock = "^5.1.0"
fakeredis = "^2.23.4"
freezegun = "^1.5.1"
pytest-cov = "^5.0.0"
pytest-benchmark = "^4.0.0"
pipdeptree = "^2.23.1"
ruff = "^0.5.5"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
python-semantic-release = "^9.8.6"
pymysql = "^1.1.1"
pillow = "^10.4.0"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Enable preview features.
preview = true

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
